@@ -5,25 +5,23 @@
   "dependencies": {
     "@emotion/react": "^11.10.6",
     "@emotion/styled": "^11.10.6",
-    "@material-ui/core": "^4.11.2",
-    "@material-ui/icons": "^4.11.3",
-    "@mui/icons-material": "^5.11.11",
+    "@mui/icons-material-material": "^5.11.16",
     "@mui/material": "^5.11.16",
+    "@mui/styles": "^5.11.16",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.22",
-    "@types/react": "^17.0.31",
+    "@types/react": "^18.0.31",
     "@types/react-dom": "^18.0.11",
-    "@types/react-router-dom": "^5.3.3",
     "axios": "^1.3.4",
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0",
-    "react-hook-form": "^7.43.8",
-    "react-router-dom": "^6.9.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.10.0",
+    "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
-  },
-  "devDependencies": {
-    "react-scripts": "5.0.1"
   },
   "scripts": {
     "start": "react-scripts start",
