@@ -1,9 +1,11 @@
 import React from 'react';
 import ContactForm from '../components/ContactForm';
-import { Box, Container, Typography, makeStyles } from '@material-ui/core';
+import { Box, Container, Typography } from '@mui/material';
 import SocialsBox from '../components/SocialsBox';
+import { makeStyles } from '@mui/styles';
+import { Theme } from '@mui/material/styles';
 
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles((theme: Theme) => ({
   header: {
     marginTop: theme.spacing(4),
     marginBottom: theme.spacing(4),
