@@ -1,21 +1,17 @@
 import React from 'react';
-import { Container, Grid, Typography } from '@material-ui/core';
-import { makeStyles } from '@material-ui/core/styles';
+import { Container, Grid, Typography } from '@mui/material';
+import { styled } from '@mui/system';
 import PluginCard from '../components/PluginCard';
 
-const useStyles = makeStyles((theme) => ({
-  root: {
-    flexGrow: 1,
-  },
-  sectionTitle: {
-    margin: theme.spacing(4, 0, 2),
-  },
+const Root = styled('div')(({ theme }) => ({
+  flexGrow: 1,
+}));
+
+const SectionTitle = styled(Typography)(({ theme }) => ({
+  margin: theme.spacing(4, 0, 2),
 }));
 
 const Plugins: React.FC = () => {
-  const classes = useStyles();
-
-  // You can replace this with data from your data.json file.
   const samplePluginData = [
     {
       iconUrl: 'https://via.placeholder.com/50',
@@ -32,7 +28,7 @@ const Plugins: React.FC = () => {
       title: 'Sample Plugin',
       description: 'This is a sample plugin.',
       link: 'https://example.com',
-    }, 
+    },
     {
       iconUrl: 'https://via.placeholder.com/50',
       category: 'Top Plugins',
@@ -48,7 +44,7 @@ const Plugins: React.FC = () => {
       title: 'Sample Plugin',
       description: 'This is a sample plugin.',
       link: 'https://example.com',
-    },{
+    }, {
       iconUrl: 'https://via.placeholder.com/50',
       category: 'Top Plugins',
       categoryLink: '#',
@@ -63,7 +59,7 @@ const Plugins: React.FC = () => {
       title: 'Sample Plugin',
       description: 'This is a sample plugin.',
       link: 'https://example.com',
-    }, 
+    },
     {
       iconUrl: 'https://via.placeholder.com/50',
       category: 'Top Plugins',
@@ -83,12 +79,13 @@ const Plugins: React.FC = () => {
     // Add more plugin objects here.
   ];
 
+  
   return (
-    <div className={classes.root}>
+    <Root>
       <Container>
-        <Typography variant="h4" component="h2" className={classes.sectionTitle}>
+        <SectionTitle variant="h4" >
           Top Plugins
-        </Typography>
+        </SectionTitle>
         <Grid container spacing={4}>
           {samplePluginData.map((plugin, index) => (
             <Grid item key={index} xs={12} sm={6} md={4} lg={3}>
@@ -99,9 +96,9 @@ const Plugins: React.FC = () => {
       </Container>
 
       <Container>
-        <Typography variant="h4" component="h2" className={classes.sectionTitle}>
+        <SectionTitle variant="h4">
           New Plugins
-        </Typography>
+        </SectionTitle>
         <Grid container spacing={4}>
           {samplePluginData.map((plugin, index) => (
             <Grid item key={index} xs={12} sm={6} md={4} lg={3}>
@@ -112,9 +109,9 @@ const Plugins: React.FC = () => {
       </Container>
 
       <Container>
-        <Typography variant="h4" component="h2" className={classes.sectionTitle}>
+        <SectionTitle variant="h4" >
           Featured Plugins
-        </Typography>
+        </SectionTitle>
         <Grid container spacing={4}>
           {samplePluginData.map((plugin, index) => (
             <Grid item key={index} xs={12} sm={6} md={4} lg={3}>
@@ -123,7 +120,7 @@ const Plugins: React.FC = () => {
           ))}
         </Grid>
       </Container>
-    </div>
+    </Root>
   );
 };
 
